(Refactoring, refining) - Fully functional with all features for logging to the Console. Includes unit tests and thorough internal documentation, although more would be nice. TODO: Finish LogfileAppender, and there is no slf4j API yet.

+++ src/main/java/com/ablestrategies/logger/IConfiguration.java
@@ -3,9 +3,13 @@ package com.ablestrategies.logger;
 /**
  * IConfiguration - Interface for a configuration provider<p/>
  * Expected Configuration Settings...<ul>
- * <li> jlogger.appenders.list  "ConsoleAppender"     Comma-delimited list </li>
- * <li> jlogger.console.prefix  "@X [@L] @C: "        See XxxFormatters.java </li>
- * <li> jlogger.default.level   "Warn"                See LogLevel.java </li>
+ * <li> jlogger.appenders.list   "ConsoleAppender"     Comma-delimited list </li>
+ * <li> jlogger.default.level    "Warn"                See LogLevel.java </li>
+ * <li> jlogger.console.prefix   "@X [@L] @C: "        See TextFormatter.java </li>
+ * <li> jlogger.logfile.prefix   "@X [@L] @C: "        See TextFormatter.java </li>
+ * <li> jlogger.logfile.name     "jlog"                See FileAppender.java </li>
+ * <li> jlogger.logfile.kmaxsize "100"                 See FileAppender.java </li>
+ * <li> jlogger.logfile.backups  "10"                  See FileAppender.java </li>
  * </ul>
  */
 public interface IConfiguration {
@@ -19,11 +23,11 @@ public interface IConfiguration {
     String getString(String key, String defaultValue);
 
     /**
-     * Get a named integer value from the configuration
+     * Get a named long integer value from the configuration
      * @param key lookup key
      * @param defaultValue default value to be used if the key or value is missing
      * @return the value
      */
-    int getInteger(String key, int defaultValue);
+    long getLong(String key, long defaultValue);
 
 }
